hr_payroll_backend/requirements.txt
# Requirements for HR & Payroll Backend
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
djoser>=2.2
django-cors-headers>=4.3
Pillow>=10.0
python-socketio>=5.10
python-dotenv>=1.0
